修改数据抽取缺失小数点的问题

+++ src/com/tend/servlet/DayZLNCtoBQSenderByIcGeneralB.java
@@ -454,7 +454,7 @@ public class DayZLNCtoBQSenderByIcGeneralB extends BaseDao implements Runnable {
 							insetSql.append(",");
 						}
 					} else {
-						insetSql.append(restNC.getInt(i));
+						insetSql.append(restNC.getDouble(i));
 						if (i < resultcount) {
 							insetSql.append(",");
 						}
