+++ src/com/tend/servlet/DayZLNCtoBQSenderByIcGeneralH.java
@@ -248,7 +248,7 @@ public class DayZLNCtoBQSenderByIcGeneralH extends BaseDao implements Runnable {
 							insetSql.append(",");
 						}
 					} else {
-						insetSql.append(restNC.getInt(i));
+						insetSql.append(restNC.getDouble(i));
 						if (i < resultcount) {
 							insetSql.append(",");
 						}
