+++ src/com/tend/servlet/DayZLNCtoBQSenderBySoSaleorderB.java
@@ -298,7 +298,7 @@ public class DayZLNCtoBQSenderBySoSaleorderB extends BaseDao implements
 							insetSql.append(",");
 						}
 					} else {
-						insetSql.append(restNC.getInt(i));
+						insetSql.append(restNC.getDouble(i));
 						if (i < resultcount) {
 							insetSql.append(",");
 						}
