+++ src/com/tend/servlet/ZLNCtoBQSenderBySoSaleinvoice.java
@@ -247,7 +247,7 @@ public class ZLNCtoBQSenderBySoSaleinvoice extends BaseDao implements Runnable {
 								insetSql.append(",");
 							}
 						}else{
-							insetSql.append(restNC.getInt(i));
+							insetSql.append(restNC.getDouble(i));
 							if(i<resultcount){
 								insetSql.append(",");
 							}
