+++ src/com/tend/servlet/ZLNCtoBQSenderBySoSaleorderB.java
@@ -301,7 +301,7 @@ public class ZLNCtoBQSenderBySoSaleorderB extends BaseDao implements Runnable {
 								insetSql.append(",");
 							}
 						}else{
-							insetSql.append(restNC.getInt(i));
+							insetSql.append(restNC.getDouble(i));
 							if(i<resultcount){
 								insetSql.append(",");
 							}
